Move fragile tests to a separate CI run
